Move necessaries to native_inference
