+++ src/marqo/inference/media_download_and_preprocess/image_download.py
