+++ src/marqo/inference/model_download/model_download.py
@@ -8,8 +8,8 @@ from urllib.error import HTTPError
 from tqdm import tqdm
 
 from marqo.core.exceptions import InternalError
-from marqo.core.inference.download_model_from_hf import download_model_from_hf
-from marqo.core.inference.download_model_from_s3 import (
+from marqo.inference.model_download.download_model_from_hf import download_model_from_hf
+from marqo.inference.model_download.download_model_from_s3 import (
     get_presigned_s3_url, get_s3_model_cache_filename, check_s3_model_already_exists,
     get_s3_model_absolute_cache_path
 )
