+++ src/marqo/inference/native_inference/embedding_models/abstract_clip_model.py
@@ -4,11 +4,9 @@ import numpy as np
 import torch
 from PIL import UnidentifiedImageError
 
-from marqo.core.inference.image_download import (_is_image, format_and_load_CLIP_images,
-                                                 format_and_load_CLIP_image)
-from marqo.core.inference.embedding_models.abstract_embedding_model import AbstractEmbeddingModel
-from marqo.core.inference.image_download import (_is_image, format_and_load_CLIP_images,
-                                                                  format_and_load_CLIP_image)
+from marqo.inference.native_inference.embedding_models import AbstractEmbeddingModel
+from marqo.inference.media_download_and_preprocess.image_download import (_is_image, format_and_load_CLIP_images,
+                                                                          format_and_load_CLIP_image)
 from marqo.s2_inference.logger import get_logger
 from marqo.s2_inference.types import *
 from marqo.tensor_search.models.private_models import ModelAuth
