+++ src/marqo/inference/native_inference/embedding_models/abstract_embedding_model.py
