+++ src/marqo/inference/native_inference/embedding_models/hugging_face_model_properties.py
@@ -8,7 +8,7 @@ from huggingface_hub.utils import HfHubHTTPError
 from pydantic import Field, validator, root_validator
 
 from marqo.base_model import ImmutableBaseModel
-from marqo.core.inference.embedding_models.marqo_base_model_properties import MarqoBaseModelProperties
+from marqo.inference.native_inference.embedding_models import MarqoBaseModelProperties
 from marqo.s2_inference.configs import ModelCache
 from marqo.s2_inference.logger import get_logger
 from marqo.tensor_search.models.private_models import ModelLocation
