+++ src/marqo/inference/native_inference/embedding_models/marqo_base_model_properties.py
