+++ src/marqo/inference/native_inference/embedding_models/open_clip_model.py
@@ -8,10 +8,10 @@ from pydantic import ValidationError
 from torchvision.transforms import Compose
 
 from marqo import marqo_docs
-from marqo.core.inference.embedding_models.abstract_clip_model import AbstractCLIPModel
-from marqo.core.inference.embedding_models.hf_tokenizer import HFTokenizer
-from marqo.core.inference.embedding_models.open_clip_model_properties import OpenCLIPModelProperties, ImagePreprocessor
-from marqo.core.inference.model_download import download_model
+from marqo.inference.native_inference.embedding_models import AbstractCLIPModel
+from marqo.inference.native_inference.embedding_models import HFTokenizer
+from marqo.inference.native_inference.embedding_models import OpenCLIPModelProperties, ImagePreprocessor
+from marqo.inference.model_download.model_download import download_model
 from marqo.exceptions import InternalError
 from marqo.s2_inference.configs import ModelCache
 from marqo.s2_inference.errors import InvalidModelPropertiesError
