+++ src/marqo/inference/native_inference/embedding_models/open_clip_model_properties.py
@@ -4,7 +4,7 @@ from typing import Optional, List
 
 from pydantic import Field, root_validator
 
-from marqo.core.inference.embedding_models.marqo_base_model_properties import MarqoBaseModelProperties
+from marqo.inference.native_inference.embedding_models import MarqoBaseModelProperties
 from marqo.tensor_search.models.private_models import ModelLocation
