+++ src/marqo/s2_inference/model_registry.py
@@ -1,9 +1,9 @@
-from marqo.core.inference.embedding_models.hugging_face_stella_model import HuggingFaceStellaModel
-from marqo.core.inference.embedding_models.languagebind_model import LanguagebindModel
+from marqo.inference.native_inference.embedding_models import HuggingFaceStellaModel
+from marqo.inference.native_inference.embedding_models import LanguagebindModel
 from marqo.s2_inference.clip_utils import CLIP, MULTILINGUAL_CLIP, FP16_CLIP, \
     get_multilingual_clip_properties
-from marqo.core.inference.embedding_models.open_clip_model import OPEN_CLIP
-from marqo.core.inference.embedding_models.hugging_face_model import HuggingFaceModel
+from marqo.inference.native_inference.embedding_models import OPEN_CLIP
+from marqo.inference.native_inference.embedding_models import HuggingFaceModel
 from marqo.s2_inference.onnx_clip_utils import CLIP_ONNX
 from marqo.s2_inference.random_utils import Random
 from marqo.s2_inference.sbert_onnx_utils import SBERT_ONNX
