+++ tests/integ_tests/core/inference/embedding_models/test_hugging_face_model.py
@@ -3,8 +3,8 @@ from unittest import mock
 
 import numpy as np
 
-from marqo.core.inference.embedding_models.hugging_face_model import HuggingFaceModel
-from marqo.core.inference.embedding_models.hugging_face_model_properties import PoolingMethod, HuggingFaceModelFlags, \
+from marqo.inference.native_inference.embedding_models import HuggingFaceModel
+from marqo.inference.native_inference.embedding_models import PoolingMethod, HuggingFaceModelFlags, \
   HuggingFaceTokenizerFlags
 from marqo.s2_inference.errors import InvalidModelPropertiesError
