+++ tests/integ_tests/core/inference/embedding_models/test_hugging_face_stella_model.py
@@ -1,6 +1,6 @@
 import unittest
 
-from marqo.core.inference.embedding_models.hugging_face_stella_model import HuggingFaceStellaModel
+from marqo.inference.native_inference.embedding_models import HuggingFaceStellaModel
 from marqo.s2_inference.errors import InvalidModelPropertiesError
