+++ tests/integ_tests/core/inference/test_languagebind_model_properties.py
@@ -3,8 +3,6 @@ import unittest
 import pytest
 from pydantic import ValidationError
 
-from marqo.core.inference.embedding_models.languagebind_model_properties import *
-
 
 @pytest.mark.unittest
 class TestLanguagebindModelProperties(unittest.TestCase):
